+++ test/unittest/CMakeLists.txt
@@ -10,8 +10,20 @@ FetchContent_Declare(
 
 FetchContent_MakeAvailable(Catch2)
 
+add_library(unittest_harness STATIC
+    tests/extent_list.cxx
+)
+
+target_include_directories(unittest_harness PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+target_compile_options(
+    unittest_harness
+    PUBLIC -std=c++17
+    PRIVATE ${COMPILER_DIAGNOSTIC_CXX_FLAGS}
+)
+
 create_test_executable(unittests
     SOURCES tests/main.cxx
+    LIBS unittest_harness
     CXX
 )
 
@@ -22,6 +34,7 @@ target_compile_options(
 )
 
 target_link_libraries(unittests Catch2::Catch2)
+target_link_libraries(unittest_harness Catch2::Catch2)
 if(ENABLE_TCMALLOC)
     target_link_libraries(unittests wt::tcmalloc)
 endif()
